+++ src/main/groovy/de/dfki/mary/voicebuilding/VoicebuildingDataPlugin.groovy
@@ -42,82 +42,38 @@ class VoicebuildingDataPlugin implements Plugin<Project> {
             project.bootstrap.dependsOn it
         }
 
-        project.task('wav')
+        project.task('wav', type: ProcessWav) {
+            dependsOn project.processDataResources
+            srcFiles = project.fileTree(project.sourceSets.data.output.resourcesDir).include('*.wav')
+            destDir = project.file("$project.buildDir/wav")
+        }
 
-        project.task('praatPitchExtractor')
+        project.task('praatPitchExtractor', type: PraatExtractPitch) {
+            dependsOn project.templates, project.wav, project.praat
+            scriptFile = project.file("$project.templates.destDir/extractPitch.praat")
+            srcFiles = project.fileTree(project.wav.destDir).include('*.wav')
+            destDir = project.file("$project.buildDir/Pitch")
+        }
 
-        project.task('praatPitchmarker')
+        project.task('praatPitchmarker', type: PraatExtractPitchmarks) {
+            dependsOn project.praatPitchExtractor, project.wav, project.praat
+            scriptFile = project.file("$project.templates.destDir/pitchmarks.praat")
+            wavFiles = project.fileTree(project.wav.destDir).include('*.wav')
+            pitchFiles = project.fileTree(project.praatPitchExtractor.destDir).include('*.Pitch')
+            destDir = project.file("$project.buildDir/PointProcess")
+        }
 
-        project.task('pitchmarkConverter')
+        project.task('pitchmarkConverter', type: PitchmarkConverter) {
+            dependsOn project.praatPitchmarker
+            srcFiles = project.fileTree(project.praatPitchmarker.destDir).include('*.PointProcess')
+            destDir = project.file("$project.buildDir/pm")
+        }
 
-        project.task('mcepExtractor')
-
-        project.voicebuilding.basenames.each { basename ->
-            project.task("${basename}_wav", type: ParallelizableExec) {
-                dependsOn project.processDataResources
-                def srcFile = project.file("$project.sourceSets.data.output.resourcesDir/${basename}.wav")
-                srcFiles << srcFile
-                destFile = project.file("$project.buildDir/wav/${basename}.wav")
-                cmd = ['sox', srcFile, destFile, 'rate', project.voice.samplingRate]
-                project.wav.dependsOn it
-            }
-
-            project.task("${basename}_extractPitch", type: PraatExec) {
-                def wavTask = project.tasks.getByName("${basename}_wav")
-                dependsOn project.templates, wavTask, project.praat
-                srcFiles << wavTask.destFile
-                destFile = project.file("$project.buildDir/pm/${basename}.Pitch")
-                scriptFile = project.file("$project.templates.destDir/extractPitch.praat")
-                project.praatPitchExtractor.dependsOn it
-                project.afterEvaluate {
-                    props = [wavFile  : wavTask.destFile,
-                             pitchFile: destFile,
-                             minPitch : (project.voice.gender == 'female') ? 100 : 75,
-                             maxPitch : (project.voice.gender == 'female') ? 500 : 300]
-                }
-            }
-
-            project.task("${basename}_extractPitchmarks", type: PraatExec) {
-                def wavTask = project.tasks.getByName("${basename}_wav")
-                def pitchTask = project.tasks.getByName("${basename}_extractPitch")
-                dependsOn project.templates, wavTask, pitchTask, project.praat
-                srcFiles << [wavTask.destFile, pitchTask.destFile]
-                destFile = project.file("$project.buildDir/pm/${basename}.PointProcess")
-                scriptFile = project.file("$project.templates.destDir/pitchmarks.praat")
-                props = [wavFile   : wavTask.destFile,
-                         pitchFile : pitchTask.destFile,
-                         pointpFile: destFile]
-                project.praatPitchmarker.dependsOn it
-            }
-
-            project.task("${basename}_convertPitchmarks", type: PitchmarkConverter) {
-                def pitchmarkTask = project.tasks.getByName("${basename}_extractPitchmarks")
-                dependsOn pitchmarkTask
-                srcFile = pitchmarkTask.destFile
-                destFile = project.file("$project.buildDir/pm/${basename}.pm")
-                project.pitchmarkConverter.dependsOn it
-            }
-
-            project.task("${basename}_mcep", type: ParallelizableExec) {
-                def wavTask = project.tasks.getByName("${basename}_wav")
-                def pitchmarkTask = project.tasks.getByName("${basename}_convertPitchmarks")
-                dependsOn wavTask, pitchmarkTask
-                srcFiles << [wavTask.destFile, pitchmarkTask.destFile]
-                destFile = project.file("$project.buildDir/mcep/${basename}.mcep")
-                cmd = ['sig2fv',
-                       '-window_type', 'hamming',
-                       '-factor', 2.5,
-                       '-otype', 'est_binary',
-                       '-coefs', 'melcep',
-                       '-melcep_order', 12,
-                       '-fbank_order', 24,
-                       '-shift', 0.01,
-                       '-preemph', 0.97,
-                       '-pm', pitchmarkTask.destFile,
-                       '-o', destFile,
-                       wavTask.destFile]
-                project.mcepExtractor.dependsOn it
-            }
+        project.task('mcepExtractor', type: ExtractMcep) {
+            dependsOn project.wav, project.pitchmarkConverter
+            wavFiles = project.fileTree(project.wav.destDir).include('*.wav')
+            pmFiles = project.fileTree(project.praatPitchmarker.destDir).include('*.pm')
+            destDir = project.file("$project.buildDir/mcep")
         }
 
         project.task('generateAllophones', type: MaryInterfaceBatchTask) {